Shift click shouldn't have been abstract.
