+++ src/main/java/org/spout/vanilla/plugin/inventory/player/PlayerCraftingInventory.java
@@ -27,9 +27,7 @@
 package org.spout.vanilla.plugin.inventory.player;
 
 import org.spout.api.inventory.ItemStack;
-import org.spout.api.material.Material;
 
-import org.spout.vanilla.api.component.inventory.PlayerInventoryComponent;
 import org.spout.vanilla.api.inventory.CraftingInventory;
 
 /**
@@ -78,24 +76,5 @@ public class PlayerCraftingInventory extends CraftingInventory {
 
 	public ItemStack getOutput() {
 		return get(OUTPUT);
-	}
-
-	@Override
-	public boolean onShiftClick(int slot, PlayerInventoryComponent toInventory) {
-		ItemStack result = get(getOutputSlot());
-		Material resultMat = (result != null ? result.getMaterial() : null);
-		if (slot == getOutputSlot() + getOffset() && result != null) {
-			while (result != null && !result.isEmpty()) {
-				toInventory.add(result);
-				set(getOutputSlot(), null);
-				result = get(getOutputSlot());
-				//Changed recipes, stop!
-				if (result != null && !result.getMaterial().isMaterial(resultMat)) {
-					break;
-				}
-			}
-			return true;
-		}
-		return false;
 	}
 }
